common.prepare: get rid of useless exceptions

+++ servers/partner-ci/builders/common.prepare.sh
@@ -70,28 +70,29 @@ function smart_erase {
       fi
     done
   fi
-
-  for vm in ${vms:?}; do
-    if virsh domstate $vm | grep -e '.*shut.*' -q; then
-      if virsh destroy $vm; then
-        echo "domain destroyed succesfully"
-        virsh undefine --remove-all-storage --snapshots-metadata $vm
-      else
-        ref=$?
-        echo "there are some troubles with virt stack, restart services and recheck"
-        sudo service libvirt-bin restart
+  if [ ! -z "$vms" ]; then
+    for vm in $vms; do
+      if virsh domstate $vm | grep -e '.*shut.*' -q; then
         if virsh destroy $vm; then
           echo "domain destroyed succesfully"
           virsh undefine --remove-all-storage --snapshots-metadata $vm
         else
           ref=$?
-          echo "there are some troubles with virt stack, please check it manually "
+          echo "there are some troubles with virt stack, restart services and recheck"
+          sudo service libvirt-bin restart
+          if virsh destroy $vm; then
+            echo "domain destroyed succesfully"
+            virsh undefine --remove-all-storage --snapshots-metadata $vm
+          else
+            ref=$?
+            echo "there are some troubles with virt stack, please check it manually "
+          fi
         fi
+      else
+        virsh undefine --remove-all-storage --snapshots-metadata $vm
       fi
-    else
-      virsh undefine --remove-all-storage --snapshots-metadata $vm
-    fi
-  done
+    done
+  fi
   dos.py sync
 }
